Sudoless Shutdown (The script! WOO!!!)
Surprisingly, going after the source code of stuff eventually lead to the right path in a humane time despite of thinking the issue being rather complicated. Regardless this completed the task and went a little bit overboard. 

Though the learning experience didn't go as deep into the hacking as I had anticipated, I ended up going through quite a lot of source code and it was an interesting learning curve. I now know stuff, that I didn't know that it existed. :D